Move client retry tests
