+++ test/async/http/client.rb
@@ -14,61 +14,6 @@ require "sus/fixtures/async"
 require "sus/fixtures/async/http"
 
 describe Async::HTTP::Client do
-	class RetryClient < Async::HTTP::Client
-		def initialize(failures)
-			@pool = FakePool.new
-			@protocol = nil
-			@retries = 3
-			@scheme = "https"
-			@authority = "example.com"
-			@failures = failures
-			@chunks = []
-		end
-		
-		attr :chunks
-		
-		class FakePool
-			def acquire
-				Object.new
-			end
-			
-			def release(connection)
-			end
-		end
-		
-		def make_response(request, connection, attempt)
-			@chunks << request.body&.read
-			
-			if failure = @failures.shift
-				raise failure
-			end
-			
-			return Protocol::HTTP::Response[200, {}, ["OK"]]
-		end
-	end
-	
-	with "retries" do
-		it "rewinds idempotent request bodies after ambiguous failures" do
-			client = RetryClient.new([EOFError.new])
-			request = Protocol::HTTP::Request["PUT", "/", {}, ["Hello"]]
-			
-			response = client.call(request)
-			
-			expect(response).to be(:success?)
-			expect(client.chunks).to be == ["Hello", "Hello"]
-		end
-		
-		it "rewinds non-idempotent request bodies after refused requests" do
-			client = RetryClient.new([Protocol::HTTP::RefusedError.new("Request not processed.")])
-			request = Protocol::HTTP::Request["POST", "/", {}, ["Hello"]]
-			
-			response = client.call(request)
-			
-			expect(response).to be(:success?)
-			expect(client.chunks).to be == ["Hello", "Hello"]
-		end
-	end
-	
 	with "basic server" do
 		include Sus::Fixtures::Async::HTTP::ServerContext
